lesson04/protobufv3_demo/py_client/pyproject.toml
[project]
name = "py-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.75.1",
    "grpcio-tools>=1.75.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
